migrate routing to dicoding routing version
