+++ src/scripts/api/config.js
@@ -1,0 +1,6 @@
+const API_ROUTES = {
+  list: " https://restaurant-api.dicoding.dev/list",
+  detail: " https://restaurant-api.dicoding.dev/detail/",
+};
+
+export default API_ROUTES;
