+++ src/scripts/pages/favourite/index.js
@@ -1,6 +1,11 @@
+import GetRestaurant from "../../api/apicall";
+
 const Favourite = {
   async render() {
     return `<h1>Favourite Page</h1>`;
+  },
+  async afterRender() {
+    await GetRestaurant.list();
   },
 };
