+++ src/scripts/pages/not-found/index.js
@@ -1,3 +1,5 @@
+import GetRestaurant from "../../api/apicall";
+
 const NotFound = {
   async render() {
     return `
@@ -5,6 +7,9 @@ const NotFound = {
         <h1>Sorry the pages not found</h1>
         <custom-links text="home" href="/"></custom-links>
         </div>`;
+  },
+  async afterRender() {
+    await GetRestaurant.list();
   },
 };
